apps/api/pyproject.toml
[project]
name = "api"
version = "0.0.1"
description = "FastAPI backend for AI Assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.9",
    "supabase>=2.10.0",
    "pydantic-settings>=2.0.0",
    "google-auth>=2.0",
    "google-api-python-client>=2.100",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = []
exclude = ["*"]
